% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandwidth_sgd.R
\name{.rows_to_df}
\alias{.rows_to_df}
\title{Convert a list of history row lists to a data.frame}
\usage{
.rows_to_df(rows)
}
\description{
Convert a list of history row lists to a data.frame
}
\keyword{internal}
