% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_ID_with_single_observation.R
\name{drop_ID_with_single_observation}
\alias{drop_ID_with_single_observation}
\title{Drop the Individuals with only one observation}
\usage{
drop_ID_with_single_observation(data, ID_num)
}
\arguments{
\item{data}{The data.frame of the raw dataset}

\item{ID_num}{The data.frame includes individuals' ID and the numbers of records (always named: Count)}
}
\value{
A data.frame that all the individuals have more than one observations
}
\description{
Drop the individual with only one observation when using "within" to avoid error in merging data frame and SDF
}
