% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_logistic.R
\name{effect_to_glmmtmb_fml}
\alias{effect_to_glmmtmb_fml}
\title{Append random-effect terms to a base formula for glmmTMB}
\usage{
effect_to_glmmtmb_fml(base_formula, effect, id_col, time_col)
}
\arguments{
\item{base_formula}{A formula object (without random-effect terms).}

\item{effect}{Character scalar: one of \code{"individual"},
\code{"time"}, \code{"two-way"}, \code{"nested"}.}

\item{id_col}{Name of the individual ID column.}

\item{time_col}{Name of the time column.}
}
\value{
A formula suitable for \code{glmmTMB::glmmTMB()}.
}
\description{
Append random-effect terms to a base formula for glmmTMB
}
\keyword{internal}
