% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_sf.R
\name{extract_geometry}
\alias{extract_geometry}
\title{Extract the geometry column from an sf object}
\usage{
extract_geometry(spatial)
}
\arguments{
\item{spatial}{An `sf` object.}
}
\value{
An `sfc` geometry column.
}
\description{
Returns the `sfc` geometry column of the supplied `sf` object.
}
\keyword{internal}
