% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protect_model_with_least_individuals.R
\name{protect_model_with_least_individuals}
\alias{protect_model_with_least_individuals}
\title{Protect each Subsample with Least Individuals to Run the Regression}
\usage{
protect_model_with_least_individuals(
  data,
  ID_list,
  index,
  kernel,
  p,
  longlat,
  bw_panel
)
}
\arguments{
\item{data}{The data.frame has been washed}

\item{ID_list}{The data.frame with individuals' ID}

\item{index}{The index C("id", "time"), here "id" is always "id", but "time" is set by user}

\item{kernel}{Kernel, default "bisquare". gaussian,exponential, bisquare, tricube, boxcar (see GWmodel::gw.weight)}

\item{p}{The power of the Minkowski distance, default is 2, i.e. the Euclidean distance (see GWmodel::bw.gwr)}

\item{longlat}{If TRUE, great circle distances will be calculated (see GWmodel::bw.gwr)}

\item{bw_panel}{The  required numbers of individuals}
}
\value{
A distance to gaurantee there is enought individuals in every subsample
}
\description{
When we use random effect model and "swar" method, the model required number of individuals the same as the estimated parameters in the subsamples. Moreover, if the number of individuals in a subsample is one, then there is no difference between panel regression and time-serious regression. At least 2. Therefore, we use this function to protect the model.
}
