% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/California.R
\docType{data}
\name{California}
\alias{California}
\title{California (SpatialPolygonsDataFrame)}
\format{
A \code{sp::SpatialPolygonsDataFrame} with 'GEOID":
\describe{
\item{GEOID}{a numeric vector, fips IDs of the counties}
}
}
\usage{
data(California)
}
\description{
The counties' boundary in California
}
\examples{
data(California)
class(California)
}
\author{
Chao Li <chaoli0394@gmail.com> Shunsuke Managi <managi.s@gmail.com>
}
\keyword{datasets}
