% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/California.R
\docType{data}
\name{California}
\alias{California}
\title{California (sf)}
\format{
An \code{sf} object with 58 rows (one per county) and two columns:
\describe{
\item{GEOID}{a numeric vector, fips IDs of the counties}
\item{geometry}{sfc_MULTIPOLYGON, county boundary polygons (CRS: NAD83 longlat)}
}
}
\usage{
data(California)
}
\description{
The counties' boundary in California
}
\examples{
data(California)
class(California)
}
\author{
Chao Li <chaoli0394@gmail.com> Shunsuke Managi <managi.s@gmail.com>
}
\keyword{datasets}
